SFR-170 Initial Commit
This is the initial commit of a fetcher of records from Project Gutenberg.
This lambda gets records from the GITenberg project, using the Github
API v4 (graphQL) to get recently updated records. It then parses the source
RDF file to generate a basic object containing the metadata for each record.

Though not implemented yet, this will pass the resulting records to a Kinesis
stream that will pass them on to further steps the ResearchNow ingest pipeline