SFR-170 Add kinesis output and refactor code
This added the output to a Kinesis stream, which will be read to provide
a further step in the ingest pipeline. The data format is fairly basic
but conforms to what should be expected in terms of a final output (though
that will be validated by Avro at a later point)

This also refactors the code to leave a cleaner endpoint and a clearer split
of responsibilities in the code.

A version of this needs to be developed that can run all of the GITenberg
repositories into the Kinesis stream for an intiial ingest run