SFR-206 Simplify OCLC Classify service function
This update removes the merging and data handling aspects of the
OCLC Classify service. It changes this function to simply
retrieve Classify records from OCLC, format them to meet the SFR
data model and return them to the data stream.

This simplifies this process and removes issues regarding how and
when data is being merged at this phase