Add alembic structure and first migration
This project will use alembic to manage db migrations. The first
such migration is here in alembic/versions. This is controlled by
alembic.ini in the project root, but that is not being committed
as it contains sensitive information. Investigation is necessary
before this can be added to version control.