Initial commit of ElasticSearch Manager
This initial commit contains a skeleton for the ES manager lambda
function. At present it will only read from the source SQS queue
and return the corresponding work record.

As the core records of this project are works, this service will
only update full work records at present, even if only a field
of a sub-record has changed. This will ensure that we reflect
all changes to the public, and more fine-grained updating can
be implemented at a later time