SFR-253 ElasticSearch ORM
Create a persistence layer/ORM for the SFR ElasticSearch layer.
This incorporates a somewhat simplified version of the SFR data
model for use in the ElasticSearch index. It exposes the fields
that are useful in search/retrieval/reading. It omits some fields,
including raw JSON blobs and hashes that are not needed by the
front-end application.

This uses the __dir__ builtin to define which fields should be
retrieved from each model. Specific relevant date fields are parsed
into explicit fields in their parent records for easier retrieval.