Fix deployment from travis build
When closing a pull request, travis should be able to build and
deploy the application. However, it was failing to derive the
proper AWS credentials and other encrypted variables for the
lambda function. This should grab the AWS credentials from
wherever they are available and include additional encrypted
variables in the travis config