Backup parent work matching for instance records
When merging instances into an existing working the lambda
attempts to locate the instance by identifiers and potentially
push fields to the existing instance's work record. If that
relationship was missing this would throw an error.

This update explicitly sets the parent work record (if it is known,
as it is if the record is coming from OCLC Classify) and should
prevent this error from occuring.