SFR-261 Fix minor glitches in hathi record ingest
Two minor issues cropped up during development. First,
the output kinesis stream should not be hard coded, but
set in an environment var. Second, the output rights and
identifiers for Hathi records should be a) there and b) always
encoded as lists so that they can be dynamically handled by the
database manager (There should never be an instance in our data
model where there is specified one-to-one relationships, at least
not yet)