SFR-260 HathiRecord builder refactor
On receiving some readability/understandability feedback this
commit refactors the record building logic into a new method
with the HathiRecord class. This makes the flow of the function
smooth and encapsulates the work being done in a more logical and
hierichical manner. It should be easier to see both the high
level operation and the detailed methods depending on your interest.