SFR-257 Resolve item link bugs
The refactor stemming from the HathiTrust ingest process introduced
several bugs into the item creator pertaining to links. In several
places it was wrongly assumed that only one link would be
received with an item, tracking down all the places where this
assumption was made proved difficult.

This commit should rectify the last of these bugs, specifically in
the item update method. Additionally this addresses a case where
we can dynamically create items from a record with multiple links
that may refer to one or more items.