SFR-257 Query Performance improvements
Due to the high volume of requests being received from the update
of HathiTrust several ineffecient queries were recieved.

First all() was being used to check for the existence of single
records. The more appropriate (and faster) one() with some
exception blocks provides a noticable performance increase.

Second, instead of returning a full ORM object when queried,
return the row id and use that to resolve a full record if necessary