Batch processing of ElasticSearch updates
This is a second attempt at speeding up the ElasticSearch
indexing process. There is a major change in approach here that
no longer tries to trigger updates 'live' as they are updated in
postgres. Instead this is designed to run on a schedule and update
records modified in the database in the past N seconds.

This optimizes things in three ways:
1) It allows us to batch retrieve records from the db
2) It allows us to bulk import records into elastic
3) It eliminates many duplicate indexing requests when the db is
updated repeatedly (Some duplication will still occur, but it will
be drastically decreased)

There are also multiple refactorings in the ESConnection class to
make it more pythonic, including list comprehensions and others
that should provide minor performance enhancements.