SFR-266 Add API versioning
This introduces the ability to version the API endpoints for use with different
versions of the the ElasticSearch index and front end. Each version is configured
as its own set of roots and has the ability to point to different ElasticSearch
indexes. By defualt the API will load the v1 (earliest) API, though this default
behavior can easily be changed.

In this commit the v2 API is a simple demo endpoint that does not implement any
functional endpoints. If this approach is approved of, new endpoints will be
implemented there.

The one possible drawback to this is integration with the API Gateway. In the dev
environment, the main Gateway API includes a version number, meaning that the
endpoints potentially will have multiple version statements, which could be
confusing. However, it is unclear if this will be the case in production, or if
it is necessary to include this.