Modify Search API endpoints, default boolean behavior
In order to handle user query strings from the front-end application
the search API should be able to understand three main search types:
- AND query
- OR query
- Quoted string (exact) query

By default ElasticSearch uses an OR query and does not take other
parameters into account. Using the query_string parameter allows
for their use, but must be customized in order to provide proper
parsing. This updates the POST endpoint to accept full query strings,
while utilizing the `default_operator` option to set the default
query type to AND. This brings the search API into closer alignment
with most users' mental model of search.

This structure allows the three conditions to be combined in any
order/permutation, and should support a full range of basic and
advanced search options.

This update is being made to the v1 of the API, and v2 will incorporate
these changes to be compatible with the second iteration of the
ElasticSearch index.