SFR-275 Update Data Model for DOAB Records
This extends the SFR data model to handle several fields being
received from the Directory of Open Access Books. This includes:
- An abstract `Language` class that allows for many languages to
be associated with `Work` and `Instance` records.
- A summary field on `Work` records that provides full-text
descriptions of records
- A new class for DOAB identifiers

The `Language` class allows SFR to store ISO 639-1/2 language codes
that improve both the retrievability and cardinality of language
data.