SFR-258 Refactor identifier cleaner as private method
Implementing the identifier cleaner at only the insert point
created a bug where the db manager _searched_ for a raw identifier
but _stored_ a cleaned version of it, which causes duplicate
identifiers to be generated in the database.

This refactors the cleaning code into a private method in the
`Identifier` class that is utilized whenever an identifier
is looked up or stored, ensuring that only clean versions are
persisted in the datbase.