SFR-258 DOAB Harvester Initial Commit
This is the initial version of a DOAB harvester function for the
SFR data pipeline. It is run on a set schedule to gather records
updated in DOAB in the last `n` days, accessing the data in DOAB's
published OAI-PMH feed.

These records are transformed into the SFR datamodel as part of the
SFR ETL pipeline and placed in the pipeline alongside records from
other sources