SFR-291 Implement search for new ElasticSearch Index
This implements a set of v2 search endpoints that utilize the new
ElasticSearch index containing updated records in the new data
model. Two endpoints are currently implemented, which both serve
GET and POST requests.

First, the v2/sfr/search endpoint supports basic/simple search,
and is designed to enable the current search options in the SFR
front end application. Queries are made by supplying a `field` and
`query`. Current fields supported are `title`, `author`, `subject`
and `keyword`, which is the default.

Second, the v2/sfr/work endpoint returns single work records
identified by the provieded identifier. This can be a UUID or other
identifier. However, if an identifier refers to more than one work
record an error will be raised. This endpoint is used to retrieve
single records for detail pages