SFR-326 Add Redis Cache to database manager
As part of the record enhancement pipeline the database manager
will call the OCLC Classify service to retrieve further metadata.
In certain instances (periodicals and monographs with many copies)
this can result in duplicate/unnecessary queries.

This update puts a Redis cache in front of the Kinesis stream that
feeds the OCLC Classify Lambda function. If a title/author or
identifier has been queried in the past 24 hours (UTC) a new query
will not be made and it will be assumed that the SFR record for the
work contains the most up to date metadata.