Improve edition fetching from Classify
This makes two changes to the main Classify query. First, it
implements a "sanity" check against works returned from the API to
ensure that collections are not returned if we are searching for
an individual work.

Second, it increases the number of editions that are returned per
request from 25 to 250, which should increase the completeness of
records in SFR.