SFR-326 OCLC Classify Performance Improvements
This implements several performance improvements for the Classify
lambda function, mainly by implementing caching via Redis and
transitioning the main output to the Lookup function to an SQS
queue.

With the redis caching, repeated requests for an identifier will be
intercepted, both to the Classify service and the lookup function.
This has the added benefit of catching any self-referencing OCLC
Classify records, which could hypothetically induce an endless loop
with the current structure.

The SQS queue will greatly speed up the rate at which the lookup
lambda can be invoked and therefore generally reduce the amount
of processing time for a single record.