SFR-311 Data parsing improvements for daily updates
Implements a daily update structure for retrieving TSV files from
the HathiFiles directory on the HathiTrust site. The updates are
published daily as a list of JSON files. This implements a method
that identifies the most recent file (which will be a gzipped TXT
file). That file is downloaded, extracted and then parsed as a TSV.

That TSV file is then parsed, with in copyright files being removed,
and the remaining rows pased to the HathiRecord class for parsing.

This implements better unit test coverage for this process.