SFR-321 Add Grouped Record Matching
In the first iterations of the SFR database manager function,
matches between records was made on the basis of finding unique
identifiers to identify matching records. This implements a
grouped set of matching that calculates the score based off the full
set of identifiers that are available in the record. The highest
scoring match is returned, and others found are returned as
potential matches to be evaluated later (data which is persisted in a
new `equivalency` table).

This should reduce the number of false negatives and allow for tighter
grouping of editions by allowing for better matching on semi-ambigious
identifiers that can be used together to uniquely identify records.