SFR-350 Add VIAF/LCNAF lookup options to search endpoint
This adds the VIAF & LCNAF authority files as lookup options for
agents. This allows precise discovery through use of a unique
identifier for agents (individuals/organizations). This will be
used for agent lookups from work detail pages, and can be documented
as an option for user queries outside of that (though it is not
currently used elsewhere in the SFR React application)