Refactor Item class and add test suite
Refactors `Item` class along same lines as `Work` and `Instance`
classes to conform to new function design. This removes functions
and makes a more maintainable class.

In addition this adds a new test suite for the class that fully
tests the methods. There is one `try/catch` block that seems unecessary
as the potentially problematic case that is caught earlier, but
leaving it in place for now in case I'm missing a potential issue.