Finalizing db-manager refactor
To support a better data pipeline structure, this completes a refactor
of the database manager function to only process record inserts. This
will help create a CRUD structure with one Lambda function taking
responsibility for each type of operation (the destroy function is yet
to be implemented)

This significantly streamlines the logic in the function and should
provide a performance boost. This is achieved partly by leveraging
some of the logic into the AWS environment itself, by allowing us to
assert that records in specific streams are create or update only.

As part of this significant additions to the test coverage for this
function have been added and general refactoring makes this function
much more testable. Further improvements can be made, but this will
be done as the performance is monitored in the dev environment.