Finalize insert-only improvements
This moves the `language` insert method which was improperly
nested.

It also moves a `flush` call from the main Instance `insert` method
to the item `createOrInsert` method, which is where it is actually
required. We only need to flush the current session if we are
storing an epub file, otherwise the process will execute normally
without needing to pass an instance identifier. These extraneous
flush invokations can signficantly slow the processing steps and
this should have a marked improvement on db processing speeds.