Improve output and identifier matching
This improves fetching of records within the session, specifically
by invoking `session.add()` at an earlier step in the function to
ensure that associated records are autoflushed and discoverable
by future queries.

This also fixes two bugs. One with putting `dict`s into the update
kinesis stream and a follow on error to language processing