Initial POC Commit
This is an initial proof-of-concept version of a Lambda function
that will take over responsibility for updating existing postgres
records from the sfr-db-manager function.

This split will bring the Lambda pipeline into better alignment with
a division of responsibilities that maps with the CRUD operations
and should provide a performance boost by simplifying the logic
through which both functions are invoked.

This shouldn't pose any issues, and barring any major complications
work will begin to refactor/refine this code further to make it
more maintainable.