Update outputManager to handle multiple shards
This minor adjustment allows for the Classify function to write
results to multiple shards of a database update Kinesis stream by
removing the hardcoded "0" and replacing it with the value of
the first assigned identifier, generally the appropriate oclc
identifier. This provides a reasonable source for the hash applied
by Kinesis and should help group records together logically.