Change Classify output to SQS
This change replaces a Kinesis stream with a queue of queries for
the OCLC Classify service. This enables a higher throughput for
the enhancement stream (should scale in line with updates made to
the source function) and enables better management in the pipeline
flow. At present, the Kinesis stream can become out-of-sync with
the database manager and SQS provides additional tools for managing
delivery delays and visibility for these messages.