Switch event source to SQS
This modifies the Classify function from an Kinesis stream event
source to an SQS source. This increases the throughput and enables
for greater control on message timeouts and delivery times.

Overall this should help manage the high volume of records that
the function will receive and allow the pipeline to scale in
proportion with the number of sources. To do so with Kinesis streams
would mean provisioning further shards, while an SQS queue should
manage this without an issue.