SQS Output bugfixes
This corrects several issues with utilizing SQS as an event source
and an output. The function was not properly configured to output
to multiple queues and this update rectifies the issue by specififying
the queue being access in the source method as opposed to the `putQueue`
method itself.

This also corrects a bug in using a queue as an event source and
allows both Kinesis and SQS sources to be configured (concurrently,
if desired)