Refactor DB model with `set`s
This implements a function-wide refactor that replaces lists with
sets, which make the management of potentially duplicate relationsips
much easier and eliminates a large amount of unecessary logic in the
control flow of the function.

Additionally, several redundant methods have been removed as a result
and the function should be able to better parse records. While this
has not been benchmarked as of yet, in several places the codebase
has been condensed significantly, which should reduce processing
times.