SFR-367 Convert app to poller instead of API
The previous iteration of this tool was configured as an API to
respond to request to generate accessibility reports. This proved
to be very difficult to scale, because even when it was run in
an asynchronous mode it would still fail to acknowledge requests
and cause a cascade of errors in the ingest pipeline.

This version instead utilizes an SQS queue to process records,
allowing it to be entirely decoupled from the function(s) that
invoke it. This involved swapping the `Express` application for
an application based on the `sqs-consumer` library, which polls
a provided queue for new records.

The internal logic of this service remains largely the same, however
some additional logic is necessary because almost no ePub files will
fit in the 256 kB limit of SQS messages. Instead a S3 key is passed
in the message for the ePub to be rated. This file is then pulled
from the ePub bucket and temporarily stored while it is scored.

This commit includes all of the new logic for process the reports and
placing them in a Kinesis stream for processing. Another commit
will include updates to the internal documentation, README and test
suite.