Speed and bug fixes
The initial commits for this branch were made with reference to
HathiTrust data and several cases were missed. These updates handle
issues found from the daily Gutenberg and DOAB ingest runs.

Specifically a case was implemented for sharding purposes in the
`outputManager` and a bug in language handling for `Work`s was
corrected.

Additionally this adds `TRANSACTION` based inserts for each
work record. This allows for more graceful rollbacks on error states
and prevents cases where a single error causes a batch of records
to be discarded and therefore a cascade of errors further on in the
pipeline.