SFR-369 Implement paging
To implement a full paging solution for the search API, several
limitations of the ElasticSearch paging structure have to be
worked around. The most significant limitation is that standard
`from/size` paging will only work up through 10,000 results by
default. In addition, paging requests can be expensive as the
position in the result set increases. To address this several
different modes of result page fetching have been implemented:

- For requests up to 10,000 records deep standard paging is
implemented using from/size
- For a request for results within 10,000 records from the end
of the result set. The applied `sort` is reversed, the desired
page retrieved and then has its results reversed again to match
the original sort order
- To retrieve `previous` and `next` pages, the `search_after`
functionality is used. This is a feature of ElasticSearch that allows
for quick retrieval of adjacent results utilizing a `sort` object
that exists on each search result
- To retrieve arbitrary pages that exist in large result sets between
the 10,000 records at either end, a staged query is made. The intial
set, up to the max of 10,000 is retrieved, and from the last record a
series of `search_after` requests are made until the desired page is
in range and can be returned. For performance these "invisible" queries
are made without returning the `_source` object and as a result perform
fairly quickly. For an especially deep paging request (say for 50k-100k
in a result set) this can take up to 5 seconds, but these requests
should be exceedingly rare. If they are not, this will have to be
re-thought, perhaps using the ElasticSearch `scroll` API

To implement these features the search methods have been turned into a
class `Search` that allows for easier construction of the ElasticSearch
query body. The main logic for constructing the query remains, with
additional helpers to manage paging and other features.