SFR-392 Initial Commit
A simple AWS Lambda function for querying/returning VIAF and LCNAF
numbers for agent names. This function is used to both normalize
name formats and to find controlled authority numbers for agents
being indexed in the SFR database.

The goal of this function is to provide faster lookups than using
fuzzy-string matching through PostgreSQL extensions and, at the
same time, enrich existing data with a more robust set of data from
the VIAF authority file.

This is achieved through a cached lookup to the OCLC VIAF service,
currently in use is the AutoSuggest endpoint of the API, which
returns only the basic fields needed for this query. Once found, the
response is cached in Redis and used to speed further queries for
the name.

The overall execution time of this function, including API latency,
must be faster than the roughly 400ms execution time of the fuzzy
string lookup in the database layer. At present, raw lookups take
between 20ms (cached) and 200ms (non-cached), which should bode
well for overall execution.