SFR-392 Reverse flow of cache check method
The `cacheCheck` method was returning early on success, which is
counter-intuitive. It was also evaluating a ternary expression
which was unnecessary, as a direct check on the length of the
returned object `keys()` allows for evaluation of the same test.

The reorganized code is therefore simpler and more logically
organized.