SFR-394 Integrate VIAF lookup
This adds the VIAF lookup lambda as the default option for finding
`agent` matches. It should reduce the reliance on the `jaro_winkler`
queries, which should be an improvement both in terms of the number
of matches it is possible make as well the speed at which these
matches are made.

The `lookupAgent` method has been refactored to invoke other methods
to execute the specific types of lookup. If a VIAF or LCNAF number
is found through a call to the service, the db is queried for the
identifier. Whatever the case, if a match cannot be found, the
function will execute a jaro_winkler query to check for any matching
records.