SFR-394 Integrate VIAF lookup service
Add an integration into the VIAF lookup service (via API call).
This adds another option for matching and merging agent records
beyond the `jaro_winkler` fuzzy text matching. By moving this
to an API, agent matching should be faster with less load on the
db instance CPU, as well as be able to take advantage of the
improved accurracy of matching through VIAF/LCNAF identifiers
in a greater number of cases.