SFR-390 Add prev/next paging to response object
Adds `prev_page_sort` and `next_page_sort` objects to the API
response object in the `paging` object. This makes it easier for
users/applications to get the previous and next pages using the
ElasticSearch `search_after` deep paging functionality.

This is not strictly necessary as the search API can understand
deep paging requests in the form of page/per_page but using these
parameters can result in faster queries for pages deeply within
a large result set.