Bug Fix - Mis-matched agent records
This implements a bug fix for unparsed HathiTrust contributor codes
that we mistakenly retrieving other agent names through the VIAF
service and resulting in deadlocks from the updates to these records.

This implements fixes from two directions. First by ensuring that
VIAF searches are not made on ambigious records. And second, by
ensuring that db entries are only updated when necessary.