Bugfix - Implement fix to ambigious Hathi identifiers
This corrects a bug in the db update process that was fetching
incorrect agent records due to an issue in data retrieval from
HathiTrust. The extra agent records were causing db deadlock issues
from repeated and unecessary update requests.

This addresses the issue by ensuring that similar requests are not
made and that unecessary updates are not issued to the database.