Bubfixes - Catch errors in ingest process
This implements two bug fixes for the HathiTrust ingest. First,
it catches a `EOFError` thrown when a pipe is closed earlier than
expected. The thread is still `join`ed and the closed, ensuring
that we do not unecessarily replicate records.

Second, implemented updates to the contributing organization
dictionary to help with lookups and prevent erronous matches further
down the ingest pipeline