Move sqlalchemy ORM into separate repo
This commit moves the shared ORM (used by several lambda functions)
into a separate repository that is loaded as a package, increasing
code reusability and streamlining the maintance/ugrade cycle for the
SFR database.

The package is deployed as a Layer, which makes for easy versioning
and allows for easier deployment of lambda functions as they no
longer need to be deployed from an Amazon Linux image.