SFR-383 Add blacklisted agent roles to search API
This commit adds an array of blacklisted agent roles to the search API, specifically the `author` endpoint. These roles (such as `publisher` and `imprint`) are associated withthe production of the book, not its writing/creation. Users generally do not expect to see results for a publisher of a book if they've executed an author search, and this prevents that confusion.

This treats the concept of `author` in a fairly broad way. It includes agents with relationships such as `illustrator` and `editor` as `author`s as they contribute to the intellectual content of a work. If further terms need to be added to the blacklist this can be easily done, but it seems best to leave this is a general sense for the time being.

This was implemented using a nested `must_not` query, which adds some complexity to the `author` endpoint, but hopefully the relatively straightforward structure of the ElasticSearch query object makes this clear.