Link Flag Update
This updates the link generation method for this function to produce a flag object for the `Link` class as opposed to a text `rel_type` granting greater flexibility to the designations for these links. Specifically for DOAB, this adds an additional method to ensure that 302 and other redirect URLs are followed to get direct access to hosted resources. Future work may involve scraping downloadable files from these URLs, but at present this provides a greater degree of access than previously