SFR-461 Add Edition range for work results
This generates a string for each work result in a search response stored in the `edition_range` field which represents the full date range for the editions present within the work. If edition publication dates could not be found it will return '????' for a date value.

This is implemented as a custom filter function and assumes that all publication dates should be represented as GMT, as they have no times associated with them.