SFR-480 Enable years filter
This fully implements the `years` filter on publication dates. There was a stub part of the `appFilters` method that was not completely implemented for the purposes of this application. This adds the ability to accurately filter results based off a range of years. This accepts a `start` year, `end` year or both, allowing users to get things before, after, or between their supplied year(s). Similar to the edition year range parsing, this enforces GMT time for all provided years to ensure accurate results are returned.

This also adds unit test for the `addFilters` method as well as integration tests for the swagger documentation (verifying their validity and version number)