Update docker image
This repo previously built the AWS Layer in the `lambci/lambda:python3.7` docker image, which was a "close approximation" of the AWS Lambda environmen. Unfortunately this included several libraries that prevent the `psycopg2` binaries from compiling properly (likely due to some installed C/C++ libraries). This has been replaced with an official `amazon/linux:latest` image which allows the binaries to compile properly for the Lambda environment.

Because this is a simpler image it is necessary to load more dependencies locally, specifically:

- python3
- git
- zip

With these packages installed the process does not need to be altered further. In the future the Lambda Python runtimes may be migrated to the amazon/linux2 image but at present that does not appear to be in the roadmap. There are pending updates to the Lambda environment and those should have an eye kept on then in case breaking changes are introduced.