Fix minor bugs in model parsing
These handle a range of minor (but important) bug fixes to how records are parsed. This is a side effect of merging the three different ORM models and is good proof of why this was a good idea, it allows us to normalize all of these methods to ensure that all functions interact with the database in the same way. These changes are:

- Ensuring that `Agent` records have access to the `tmp_` fields when they've been retrieved from the database
- Fix the `lookup` methods form `Instance` and `Item` records
- Fix `alt_title` references in several places