Resolve issue with updated Agents not being added to new Works
`Agent` records were not being added to newly created `Work` records if they already existed in the database. This was because the model was failing to properly parse their roles, which was leaving an empty array for parsing and the model unable to assign the relationship. This fix ensures that all roles are properly assigned and makes the `__init__` method for each of the relational tables explicit. This should have no material impact, but should make this easier to understand/maintain.