Catch deadlock and retry
With multiple processes updating the database concurrently there exists the possibility that a single record will be updated at once (generally this can happen with agents and other entries that share many relationships to the core record types) This catches the raised errors and retries the record until it succeeds. There might need to be a break after a certain number of retries, but hypothetically this should always succeed eventually.