Move edition_range value to within the _source object
This moves the calculated `edition_range` value to within the returned `_source` object. This object contains all of the metadata contained by a `Work`. Prior to this the `edition_range` values were being included in the main `hit` body, which generally contains metadata describing the `Work` record and as a result was confusing semantically. This should provide a more logical place to find this data and make for easier parsing of results in the front-end application