Improve sort_title parsing for work records
In order to properly sort in the front-end application, work titles must have stopwords (a, an, the) removed from the start of the string as well as be converted to lower case. This change implements these edits and adds stopwords for French and Spanish as well, checking for the language of the work to determine the list to use.