Fix oversize JSON response
Currently the search API returns a full set of editions for the search result response. This is unecessary as only the top three will ever be displayed. Too large responses both affect performance and can cause axios to throw a 500 error when handling a response of over 5MB (roughly).

This also implements a stub function that will be able to sort/filter editions depending on the current search criteria, which is a task that is currently in the backlog. This implements no functionality, but provides a method and testing framework that can be built into this.