Fix empty author/title string parsing
This function was parsing empty strings as valid authors and titles for the purposes of querying the Classify service. However, Classify treats an empty string as no value, which was returning erronous results (Far too many matches for generic titles usually) A previous fix was intended to prevent this, but failed to catch these cases (only catching cases of `None`).

An additional test was added to ensure that empty strings throw an error