SFR-521 Ensure that all buckets are created for language facet
Previously only the top 10 buckets were being created for the language facet, this implements the `size` parameter to ensure that all are returned. For most facets this type of fixed size would be inadvisable, but since there are a limited number of languages, defined by ISO-639 this can be safely set.