SFR-521 Implement different date range checking
Previously the date range filter checked for an INTERSECTION of the date range for editions with the provided date range. This changes this setting to use WITHIN, which works more in line with the intended design of the date filter.

The only drawback to this is that for fuzzy or uncertain dates this will cause them to be dropped if part of their range falls outside of the range being searched.