SFR-80 Add publication date sort option
Adds publication date as attached to `Work` `Instances` as a sort method. The involves two `date` fields: `pub_date_sort` and `pub_date_sort_desc` which are necessary as the `DateRange` field used for searching/filtering cannot be used to sort due to ElasticSearch constraints.

These fields are used to sort in two directions: the `desc` field for new=>old sorts and the plain sort field for old=>new. This means that the sort order can be different if a work has a wide range of publication dates, though the current assumption is that these differences will match the mental model of the application's users.

This includes further unit tests