SFR-80 Add date sort documentation
This adds documentation to the `date` sort option.

Additionally this extends the dev environment for the API by adding a new piece of middleware to the express app (for dev use only) which renders the Swagger documentation when running locally. This allows the documentation output to be more easily tested/verified before deployment. This integrates with the swagger integration tests to provide both an automated and manual review process for this documentation