SFR-515 Implement edition validity filtering
This resolves the bug raised in the (now sub-task) SFR-533 ticket. It uses the same criteria as applied in the front end to remove invalid edition records that are missing metadata. This also involves tweaks to the `edition_range` calculations to ensure that they are made on a deep copy of the edition array and not on the filtered version kept within the `hit` object itself