SFR-515 Update filter options to work with edition filtering
Previously the `filters` were applied seperately as nested queries. This worked smoothly while they were not being applied to the editions. With the use of `inner_hits` this method needed to be updated to apply any and all filters _together_ on the editions. This combines the filter options into a singled `nested` query that applied `bool` options if multiple filters of a single type are applied.

This involves some dynamic generation of the filter options, however it must be somewhat fixed in order to generated the deeply nested ElasticSearch query document. This adds some overhead to future development, but is necessary to provide this functionality