SFR-525 Block empty date filters
It is possible to receive an "empty" date filter as part of a search request from the front-end app (with empty strings for the `start` and `end` years), which inadvertently filters the results for only those with valid publication dates. This may not be desirable and when "clearing" filters can result in discrepancies in the total number of results returned.

This adds a simple check to be sure that the filter contains at least one valid year, indicating that this is a legitimate filter request