SFR-288 Add multiple query option to search
This adds the option for users to pass multiple `query` objects as search parameters to the main search endpoint. This enables a multi-field advanced search option. This takes the form of a `queries` array of objects that must contain `field` and `query` fields, the same as the single query search option.

At present these terms are joined with AND, though this will be made user-configurable later on.

Minor refactoring of the `Search` class took place to enable this option and return logical error messages in the case of missing parameters.