SFR-47 Add Utility endpoint with languages option
This adds a new route and endpoint to the ResearchNow API. The `utils` route will include supporting services that do not provide core search/lookup services. At present the only `util` is a `GET` only endpoint `languages` which returns an unique array of languages in the database. This can be used to provide search form options, it is intended to support a dropdown/type-ahead for the ResearchNow advanced search page.

Optionally the `languages` endpoint accepts a `total` parameter that if set to `true` will return the total number of `work` records associated with each distinct language. This potentially supports the gathering of interesting statistics on ResearchNow and supporting of features such as counts of records displayed on the project's homepage (as has been previously discussed)